Feat: #4 [snsproject] 안드로이드 기본 세팅
